+++ derive/src/args.rs
@@ -151,6 +151,10 @@ pub struct SimpleObjectField {
     #[darling(default)]
     pub requires: Option<String>,
     #[darling(default)]
+    pub shareable: bool,
+    #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
     pub guard: Option<SpannedValue<String>>,
     #[darling(default)]
     pub visible: Option<Visible>,
@@ -198,6 +202,10 @@ pub struct SimpleObject {
     #[darling(default)]
     pub extends: bool,
     #[darling(default)]
+    pub shareable: bool,
+    #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default, multiple, rename = "concrete")]
     pub concretes: Vec<ConcreteType>,
@@ -222,6 +230,7 @@ pub struct Argument {
     pub process_with: Option<String>,
     pub key: bool, // for entity
     pub visible: Option<Visible>,
+    pub inaccessible: bool,
     pub secret: bool,
 }
 
@@ -235,6 +244,8 @@ pub struct Object {
     pub rename_args: Option<RenameRule>,
     pub cache_control: CacheControl,
     pub extends: bool,
+    pub shareable: bool,
+    pub inaccessible: bool,
     pub use_type_description: bool,
     pub visible: Option<Visible>,
     pub serial: bool,
@@ -278,6 +289,8 @@ pub struct ObjectField {
     pub external: bool,
     pub provides: Option<String>,
     pub requires: Option<String>,
+    pub shareable: bool,
+    pub inaccessible: bool,
     pub guard: Option<SpannedValue<String>>,
     pub visible: Option<Visible>,
     pub complexity: Option<ComplexityType>,
@@ -315,6 +328,8 @@ pub struct Enum {
     pub remote: Option<String>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
 }
 
 #[derive(FromVariant)]
@@ -330,6 +345,8 @@ pub struct EnumItem {
     pub deprecation: Deprecation,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
 }
 
 #[derive(FromDeriveInput)]
@@ -346,6 +363,8 @@ pub struct Union {
     pub name: Option<String>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
 }
 
 #[derive(FromVariant)]
@@ -388,6 +407,8 @@ pub struct InputObjectField {
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
     pub secret: bool,
 }
 
@@ -409,6 +430,8 @@ pub struct InputObject {
     pub rename_fields: Option<RenameRule>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
     #[darling(default, multiple, rename = "concrete")]
     pub concretes: Vec<ConcreteType>,
     // for SimpleObject
@@ -430,6 +453,8 @@ pub struct OneofObjectField {
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
     pub secret: bool,
 }
 
@@ -449,6 +474,8 @@ pub struct OneofObject {
     pub rename_fields: Option<RenameRule>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
     #[darling(default, multiple, rename = "concrete")]
     pub concretes: Vec<ConcreteType>,
 }
@@ -466,6 +493,8 @@ pub struct InterfaceFieldArgument {
     pub default_with: Option<LitStr>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
     #[darling(default)]
     pub secret: bool,
 }
@@ -491,6 +520,10 @@ pub struct InterfaceField {
     pub requires: Option<String>,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
+    pub shareable: bool,
 }
 
 #[derive(FromVariant)]
@@ -521,6 +554,8 @@ pub struct Interface {
     pub extends: bool,
     #[darling(default)]
     pub visible: Option<Visible>,
+    #[darling(default)]
+    pub inaccessible: bool,
 }
 
 #[derive(FromMeta, Default)]
@@ -530,6 +565,7 @@ pub struct Scalar {
     pub name: Option<String>,
     pub use_type_description: bool,
     pub visible: Option<Visible>,
+    pub inaccessible: bool,
     pub specified_by_url: Option<String>,
 }
 
@@ -593,6 +629,10 @@ pub struct MergedObject {
     pub cache_control: CacheControl,
     #[darling(default)]
     pub extends: bool,
+    #[darling(default)]
+    pub shareable: bool,
+    #[darling(default)]
+    pub inaccessible: bool,
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
@@ -741,6 +781,8 @@ pub struct NewType {
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
+    pub inaccessible: bool,
+    #[darling(default)]
     pub specified_by_url: Option<String>,
 }
 
@@ -764,6 +806,8 @@ pub struct ComplexObjectField {
     pub external: bool,
     pub provides: Option<String>,
     pub requires: Option<String>,
+    pub shareable: bool,
+    pub inaccessible: bool,
     pub guard: Option<SpannedValue<String>>,
     pub visible: Option<Visible>,
     pub complexity: Option<ComplexityType>,