+++ derive/src/oneof_object.rs
@@ -16,6 +16,7 @@ pub fn generate(object_args: &args::OneofObject) -> GeneratorResult<TokenStream>
     let desc = get_rustdoc(&object_args.attrs)?
         .map(|s| quote! { ::std::option::Option::Some(#s) })
         .unwrap_or_else(|| quote! {::std::option::Option::None});
+    let inaccessible = object_args.inaccessible;
     let gql_typename = object_args
         .name
         .clone()
@@ -41,6 +42,7 @@ pub fn generate(object_args: &args::OneofObject) -> GeneratorResult<TokenStream>
                 .rename_fields
                 .rename(enum_name.to_string(), RenameTarget::Field)
         });
+        let inaccessible = variant.inaccessible;
         let desc = get_rustdoc(&object_args.attrs)?
             .map(|s| quote! { ::std::option::Option::Some(#s) })
             .unwrap_or_else(|| quote! {::std::option::Option::None});
@@ -80,6 +82,7 @@ pub fn generate(object_args: &args::OneofObject) -> GeneratorResult<TokenStream>
                     ty: <::std::option::Option<#ty> as #crate_name::InputType>::create_type_info(registry),
                     default_value: ::std::option::Option::None,
                     visible: #visible,
+                    inaccessible: #inaccessible,
                     is_secret: #secret,
                 });
             });
@@ -133,6 +136,7 @@ pub fn generate(object_args: &args::OneofObject) -> GeneratorResult<TokenStream>
                             fields
                         },
                         visible: #visible,
+                        inaccessible: #inaccessible,
                         rust_typename: ::std::any::type_name::<Self>(),
                         oneof: true,
                     })
@@ -183,6 +187,7 @@ pub fn generate(object_args: &args::OneofObject) -> GeneratorResult<TokenStream>
                             fields
                         },
                         visible: #visible,
+                        inaccessible: #inaccessible,
                         rust_typename: ::std::any::type_name::<Self>(),
                         oneof: true,
                     })