+++ derive/src/simple_object.rs
@@ -30,6 +30,8 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
     let ident = &object_args.ident;
     let (impl_generics, ty_generics, where_clause) = object_args.generics.split_for_impl();
     let extends = object_args.extends;
+    let shareable = object_args.shareable;
+    let inaccessible = object_args.inaccessible;
     let gql_typename = if !object_args.name_type {
         object_args
             .name
@@ -127,6 +129,8 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             .unwrap_or_else(|| quote! {::std::option::Option::None});
         let field_deprecation = gen_deprecation(&field.deprecation, &crate_name);
         let external = field.external;
+        let shareable = field.shareable;
+        let inaccessible = field.inaccessible;
         let requires = match &field.requires {
             Some(requires) => quote! { ::std::option::Option::Some(#requires) },
             None => quote! { ::std::option::Option::None },
@@ -174,6 +178,8 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                     external: #external,
                     provides: #provides,
                     requires: #requires,
+                    shareable: #shareable,
+                    inaccessible: #inaccessible,
                     visible: #visible,
                     compute_complexity: ::std::option::Option::None,
                 });
@@ -329,6 +335,8 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                         },
                         cache_control: #cache_control,
                         extends: #extends,
+                        shareable: #shareable,
+                        inaccessible: #inaccessible,
                         keys: ::std::option::Option::None,
                         visible: #visible,
                         is_subscription: false,
@@ -393,6 +401,8 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                         },
                         cache_control: #cache_control,
                         extends: #extends,
+                        shareable: #shareable,
+                        inaccessible: #inaccessible,
                         keys: ::std::option::Option::None,
                         visible: #visible,
                         is_subscription: false,