+++ derive/src/subscription.rs
@@ -119,6 +119,7 @@ pub fn generate(
                             ty: <#ty as #crate_name::InputType>::create_type_info(registry),
                             default_value: #schema_default,
                             visible: #visible,
+                            inaccessible: false,
                             is_secret: #secret,
                         });
                     });
@@ -267,7 +268,9 @@ pub fn generate(
                     external: false,
                     requires: ::std::option::Option::None,
                     provides: ::std::option::Option::None,
+                    shareable: false,
                     visible: #visible,
+                    inaccessible: false,
                     compute_complexity: #complexity,
                 });
             });
@@ -408,6 +411,8 @@ pub fn generate(
                     extends: #extends,
                     keys: ::std::option::Option::None,
                     visible: #visible,
+                    shareable: false,
+                    inaccessible: false,
                     is_subscription: true,
                     rust_typename: ::std::any::type_name::<Self>(),
                 })