+++ src/registry/mod.rs
@@ -111,6 +111,7 @@ pub struct MetaInputValue {
     pub ty: String,
     pub default_value: Option<String>,
     pub visible: Option<MetaVisibleFn>,
+    pub inaccessible: bool,
     pub is_secret: bool,
 }
 
@@ -166,6 +167,8 @@ pub struct MetaField {
     pub requires: Option<&'static str>,
     pub provides: Option<&'static str>,
     pub visible: Option<MetaVisibleFn>,
+    pub shareable: bool,
+    pub inaccessible: bool,
     pub compute_complexity: Option<ComplexityType>,
 }
 
@@ -175,6 +178,7 @@ pub struct MetaEnumValue {
     pub description: Option<&'static str>,
     pub deprecation: Deprecation,
     pub visible: Option<MetaVisibleFn>,
+    pub inaccessible: bool,
 }
 
 type MetaVisibleFn = fn(&Context<'_>) -> bool;
@@ -209,6 +213,7 @@ pub enum MetaType {
         description: Option<&'static str>,
         is_valid: fn(value: &Value) -> bool,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         specified_by_url: Option<&'static str>,
     },
     Object {
@@ -217,8 +222,10 @@ pub enum MetaType {
         fields: IndexMap<String, MetaField>,
         cache_control: CacheControl,
         extends: bool,
+        shareable: bool,
         keys: Option<Vec<String>>,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         is_subscription: bool,
         rust_typename: &'static str,
     },
@@ -230,6 +237,7 @@ pub enum MetaType {
         extends: bool,
         keys: Option<Vec<String>>,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         rust_typename: &'static str,
     },
     Union {
@@ -237,6 +245,7 @@ pub enum MetaType {
         description: Option<&'static str>,
         possible_types: IndexSet<String>,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         rust_typename: &'static str,
     },
     Enum {
@@ -244,6 +253,7 @@ pub enum MetaType {
         description: Option<&'static str>,
         enum_values: IndexMap<&'static str, MetaEnumValue>,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         rust_typename: &'static str,
     },
     InputObject {
@@ -251,6 +261,7 @@ pub enum MetaType {
         description: Option<&'static str>,
         input_fields: IndexMap<String, MetaInputValue>,
         visible: Option<MetaVisibleFn>,
+        inaccessible: bool,
         rust_typename: &'static str,
         oneof: bool,
     },
@@ -492,6 +503,8 @@ impl Registry {
                         fields: Default::default(),
                         cache_control: Default::default(),
                         extends: false,
+                        shareable: false,
+                        inaccessible: false,
                         keys: None,
                         visible: None,
                         is_subscription: false,
@@ -619,6 +632,8 @@ impl Registry {
                     external: false,
                     requires: None,
                     provides: None,
+                    shareable: false,
+                    inaccessible: false,
                     visible: None,
                     compute_complexity: None,
                 },
@@ -633,6 +648,7 @@ impl Registry {
                     description: None,
                     possible_types,
                     visible: None,
+                    inaccessible: false,
                     rust_typename: "async_graphql::federation::Entity",
                 },
             );
@@ -653,6 +669,7 @@ impl Registry {
                                     ty: "[_Any!]!".to_string(),
                                     default_value: None,
                                     visible: None,
+                                    inaccessible: false,
                                     is_secret: false,
                                 },
                             );
@@ -664,7 +681,9 @@ impl Registry {
                         external: false,
                         requires: None,
                         provides: None,
+                        shareable: false,
                         visible: None,
+                        inaccessible: false,
                         compute_complexity: None,
                     },
                 );
@@ -694,7 +713,9 @@ impl Registry {
                             external: false,
                             requires: None,
                             provides: None,
+                            shareable: false,
                             visible: None,
+                            inaccessible: false,
                             compute_complexity: None,
                         },
                     );
@@ -702,8 +723,10 @@ impl Registry {
                 },
                 cache_control: Default::default(),
                 extends: false,
+                shareable: false,
                 keys: None,
                 visible: None,
+                inaccessible: false,
                 is_subscription: false,
                 rust_typename: "async_graphql::federation::Service",
             },