+++ src/types/empty_mutation.rs
@@ -54,8 +54,10 @@ impl OutputType for EmptyMutation {
             fields: Default::default(),
             cache_control: Default::default(),
             extends: false,
+            shareable: false,
             keys: None,
             visible: None,
+            inaccessible: false,
             is_subscription: false,
             rust_typename: std::any::type_name::<Self>(),
         })